wip round logic, fix timers and teardown for phaser entities
